update project zhiwangxueweilunwen

+++ Project/ZhiWangLunWen/main/lunwen_data/xueWeiLunWen_lunWenData_main.py
@@ -6,8 +6,8 @@
 import sys
 import os
 import time
-import hashlib
 import traceback
+import ast
 from multiprocessing import Pool
 from multiprocessing.dummy import Pool as ThreadPool
 
@@ -32,7 +32,7 @@ class BastSpiderMain(object):
                                                                                          retry=config.RETRY,
                                                                                          proxy_country=config.COUNTRY)
         self.server = service.XueWeiLunWen_LunWenDataServer(logging=LOGGING)
-        self.dao = dao.XueWeiLunWen_LunWenDataDao(logging=LOGGING)
+        self.dao = dao.XueWeiLunWen_LunWenDataDao(logging=LOGGING, mysqlpool_number=5, redispool_number=5)
 
 
 class SpiderMain(BastSpiderMain):
@@ -40,22 +40,37 @@ class SpiderMain(BastSpiderMain):
         super().__init__()
 
     def handle(self, task_data):
+        # 将任务数据转换成python字典类型，redis返回的是str格式
+        task_data_dict = ast.literal_eval(task_data)
+
+        # 获取数据主键
+        sha = task_data_dict['sha']
         # 获取task数据
-        task = self.server.getTask(task_data)
-        # task = {'title': '高强水泥基复合材料雷达波吸收性能研究', 'zuoZhe': '国爱丽', 'daoShiXingMing': '巴恒静', 'xiaZaiCiShu': '808',
-        #         'shiJian': '2010', 'xueWeiLeiXing': '博士', 'zhuanYe': '工学_土木工程_结构工程',
-        #         'url': 'http://kns.cnki.net/kcms/detail/detail.aspx?sfield=FN&dbCode=CDFD&fileName=2011015993.nh&tableName=CDFD0911'}
-        sha = hashlib.sha1(task['url'].encode('utf-8')).hexdigest()
+        task = self.server.getTask(task_data_dict)
+
+        if task is None:
+            # 数据有异常, 从mysql数据库删除本数据, 并退出当前任务
+            self.dao.deleteLunWenUrl(sha=sha)
+            return
+
+        # # task = {'title': '高强水泥基复合材料雷达波吸收性能研究', 'zuoZhe': '国爱丽', 'daoShiXingMing': '巴恒静', 'xiaZaiCiShu': '808',
+        # #         'shiJian': '2010', 'xueWeiLeiXing': '博士', 'zhuanYe': '工学_土木工程_结构工程',
+        # #         'url': 'http://kns.cnki.net/kcms/detail/detail.aspx?sfield=FN&dbCode=CDFD&fileName=2011015993.nh&tableName=CDFD0911'}
         url = task['url']
 
         # 查询当前文章是否被抓取过
         status = self.dao.getTaskStatus(sha=sha)
+
+        # False说明没有被抓取过
         if status is False:
 
             # 获取论文主页html源码
             article_html = self.download_middleware.getResp(url=url, mode='get')
+
+            # 判断响应状态
             if article_html['status'] == 0:
                 article_html = article_html['data'].content.decode('utf-8')
+
                 # 获取发布单位
                 task['faBuDanWei'] = self.server.getFaBuDanWei(article_html)
                 # 获取摘要
@@ -119,11 +134,13 @@ class SpiderMain(BastSpiderMain):
                     self.dao.saveMediaToMysql(url=media['url'], type='image')
                 # # 记录已抓取任务
                 # self.dao.saveComplete(table=config.MYSQL_REMOVAL, sha=sha)
-                # 删除任务
-                self.dao.deleteLunWenUrl(sha=sha)
+
                 # 保存数据
-                status = self.dao.saveDataToHbase(data=task)
-                LOGGING.info(status)
+                status = ast.literal_eval(self.dao.saveDataToHbase(data=task))
+                if str(status['resultCode']) == '0':
+                    self.dao.deleteLunWenUrl(sha=sha)
+
+                LOGGING.warning(str(status))
 
             else:
                 LOGGING.error('获取论文页html源码失败，url: {}'.format(url))
@@ -131,7 +148,6 @@ class SpiderMain(BastSpiderMain):
         else:
             LOGGING.warning('{}: 已被抓取过'.format(sha))
             self.dao.deleteLunWenUrl(sha=sha)
-
 
     def start(self):
         while True:
@@ -166,6 +182,8 @@ if __name__ == '__main__':
     begin_time = time.time()
     po = Pool(config.XUEWEILUNWEN_LUNWENDATA_PROCESS_NUMBER)
     for i in range(config.XUEWEILUNWEN_LUNWENDATA_PROCESS_NUMBER):
+    # po = Pool(1)
+    # for i in range(1):
         po.apply_async(func=process_start)
     po.close()
     po.join()
